+++ source-code/core/src/config/schema.ts
@@ -1,6 +1,6 @@
 import type * as ast from "../ast/index.js"
 import type { FS } from "../fs/index.js"
-import type { LintRule } from "../lint/alternative/types.js"
+import type { LintRule } from "../lint/rule.js"
 import type { $import } from "./environment-functions/index.js"
 
 /**
