+++ source-code/core/src/lint/createLintRule.test.ts
