+++ source-code/core/src/lint/report.ts
@@ -1,15 +1,14 @@
-import type { LintedNode, LintLevel, LintReport } from "./context.js"
-import type { LintableNode, LintRuleId } from "./rule.js"
+import type { LintableNode, LintReport, LintRule, NodeWithLints } from "./types.js"
 
 /**
  * Creates a function that can be used to report lint issues.
  */
-export const createReportFunction = (_args: { id: LintRuleId; level: LintLevel }) => {
+export const createReportFunction = (_args: { id: LintRule["id"]; level: LintRule["level"] }) => {
 	return (args: { node: LintableNode; message: string }) => {
 		if (!args.node) return
 
 		args.node.lint = [
-			...((args.node as LintedNode).lint ?? []),
+			...((args.node as NodeWithLints).lint ?? []),
 			{
 				id: _args.id,
 				level: _args.level,
